add spinner animation + rename firebase config
